# BE/docs/api_doc/swagger.json
{"openapi":"3.1.0","info":{"title":"POKEMON","description":"Juegos de Cartes de POKEMON","contact":{"name":"John Doe","email":"dumy@dumy"},"version":"0.0.1-SNAPSHOT"},"servers":[{"url":"http://localhost:8080","description":"Juegos de Cartes de POKEMON"}],"tags":[{"name":"Cards","description":"Endpoints for card catalog operations"}],"paths":{"/api/decks/{id}":{"get":{"tags":["deck-controller"],"operationId":"getDeck","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeckResponse"}}}}}},"put":{"tags":["deck-controller"],"operationId":"updateDeck","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDeckRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeckResponse"}}}}}},"delete":{"tags":["deck-controller"],"operationId":"deleteDeck","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/users/register":{"post":{"tags":["user-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/players":{"get":{"tags":["guest-player-controller"],"operationId":"listPlayers","parameters":[{"name":"userId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GuestPlayerResponse"}}}}}}},"post":{"tags":["guest-player-controller"],"operationId":"createGuestPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGuestPlayerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestPlayerResponse"}}}}}}},"/api/matches":{"post":{"tags":["match-controller"],"operationId":"createMatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponse"}}}}}}},"/api/matches/{id}/join":{"post":{"tags":["match-controller"],"operationId":"joinMatch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinMatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponse"}}}}}}},"/api/matches/{id}/actions":{"post":{"tags":["game-action-controller"],"operationId":"executeAction","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameActionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameActionResponse"}}}}}}},"/api/decks":{"get":{"tags":["deck-controller"],"operationId":"listDecksByPlayer","parameters":[{"name":"playerId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeckResponse"}}}}}}},"post":{"tags":["deck-controller"],"operationId":"createDeck","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeckRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeckResponse"}}}}}}},"/api/decks/{id}/validate":{"post":{"tags":["deck-controller"],"operationId":"validateDeck","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeckValidationResponse"}}}}}}},"/api/decks/validate":{"post":{"tags":["deck-controller"],"operationId":"validateCards","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateDeckRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeckValidationResponse"}}}}}}},"/api/cards/sync":{"post":{"tags":["Cards"],"summary":"Manual card sync","description":"Triggers a manual synchronization of all cards from the Pokemon TCG API","operationId":"syncCards","responses":{"200":{"description":"Sync completed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSyncResponse"}}}},"500":{"description":"Sync failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSyncResponse"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/players/{id}":{"get":{"tags":["guest-player-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestPlayerResponse"}}}}}}},"/api/matches/{id}/state":{"get":{"tags":["match-controller"],"operationId":"getMatchState","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"playerId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchStateResponse"}}}}}}},"/api/cards":{"get":{"tags":["Cards"],"summary":"Search cards with filters","description":"Search the local card catalog with optional filters and pagination","operationId":"searchCards","parameters":[{"name":"query","in":"query","description":"Search query (name LIKE)","required":false,"schema":{"type":"string"}},{"name":"supertype","in":"query","description":"Card supertype filter (POKEMON, TRAINER, ENERGY)","required":false,"schema":{"type":"string"}},{"name":"setCode","in":"query","description":"Set code filter","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page number (0-based)","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","description":"Page size","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSearchResponse"}}}},"200":{"description":"Successful search","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSearchResponse"}}}}}}},"/api/cards/{id}":{"get":{"tags":["Cards"],"summary":"Get card by ID","description":"Returns full card details including attacks, weaknesses, and resistances","operationId":"getCardById","parameters":[{"name":"id","in":"path","description":"Card ID (e.g., xy1-1)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Card found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDetailResponse"}}}},"404":{"description":"Card not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDetailResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDetailResponse"}}}}}}}},"components":{"schemas":{"DeckCardRequest":{"properties":{"cardId":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}},"UpdateDeckRequest":{"properties":{"name":{"type":"string"},"cards":{"type":"array","items":{"$ref":"#/components/schemas/DeckCardRequest"}}}},"DeckCardResponse":{"properties":{"cardId":{"type":"string"},"name":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"supertype":{"type":"string"},"isBasicEnergy":{"type":"boolean"}}},"DeckResponse":{"properties":{"id":{"type":"string"},"name":{"type":"string"},"ownerPlayerId":{"type":"string"},"source":{"type":"string"},"totalCards":{"type":"integer","format":"int32"},"valid":{"type":"boolean"},"cards":{"type":"array","items":{"$ref":"#/components/schemas/DeckCardResponse"}},"validation":{"$ref":"#/components/schemas/DeckValidationResponse"}}},"DeckValidationError":{"properties":{"code":{"type":"string"},"message":{"type":"string"},"details":{"type":"object"}}},"DeckValidationResponse":{"properties":{"valid":{"type":"boolean"},"errors":{"type":"array","items":{"$ref":"#/components/schemas/DeckValidationError"}}}},"CreateUserRequest":{"properties":{"email":{"type":"string"},"password":{"type":"string"},"displayName":{"type":"string"}}},"GuestPlayerResponse":{"properties":{"id":{"type":"string"},"displayName":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"ownerUserId":{"type":"string"}}},"UserResponse":{"properties":{"id":{"type":"string"},"email":{"type":"string"},"displayName":{"type":"string"},"players":{"type":"array","items":{"$ref":"#/components/schemas/GuestPlayerResponse"}}}},"LoginRequest":{"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"CreateGuestPlayerRequest":{"properties":{"displayName":{"type":"string"},"userId":{"type":"string"}}},"CreateMatchRequest":{"properties":{"player1Id":{"type":"string"},"player1Name":{"type":"string"},"player1DeckId":{"type":"string"},"player2Id":{"type":"string"},"player2Name":{"type":"string"},"player2DeckId":{"type":"string"}}},"MatchResponse":{"properties":{"id":{"type":"string"},"status":{"type":"string"},"currentPhase":{"type":"string"},"turnNumber":{"type":"integer","format":"int32"},"currentPlayerId":{"type":"string"},"firstPlayerId":{"type":"string"},"winnerPlayerId":{"type":"string"},"finishReason":{"type":"string"},"players":{"type":"array","items":{"$ref":"#/components/schemas/PlayerSummary"}},"createdAt":{"type":"string","format":"date-time"}}},"PlayerSummary":{"properties":{"playerId":{"type":"string"},"side":{"type":"string"},"displayName":{"type":"string"}}},"JoinMatchRequest":{"properties":{"playerName":{"type":"string"},"deckId":{"type":"string"},"playerId":{"type":"string"}}},"GameActionRequest":{"properties":{"type":{"type":"string"},"playerId":{"type":"string"},"payload":{"type":"object"},"clientRequestId":{"type":"string"}}},"ErrorDto":{"properties":{"code":{"type":"string"},"message":{"type":"string"},"details":{"type":"object"}}},"GameActionResponse":{"properties":{"success":{"type":"boolean"},"clientRequestId":{"type":"string"},"publicState":{"type":"object"},"privateState":{"type":"object"},"events":{"type":"array","items":{"$ref":"#/components/schemas/GameEventDto"}},"error":{"$ref":"#/components/schemas/ErrorDto"}}},"GameEventDto":{"properties":{"type":{"type":"string"},"message":{"type":"string"},"payload":{"type":"object"}}},"CreateDeckRequest":{"properties":{"name":{"type":"string"},"userId":{"type":"string"},"cards":{"type":"array","items":{"$ref":"#/components/schemas/DeckCardRequest"}}}},"ValidateCardEntry":{"properties":{"cardId":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}},"ValidateDeckRequest":{"properties":{"cards":{"type":"array","items":{"$ref":"#/components/schemas/ValidateCardEntry"}}}},"CardSyncResponse":{"properties":{"success":{"type":"boolean"},"message":{"type":"string"},"newCards":{"type":"integer","format":"int32"},"updatedCards":{"type":"integer","format":"int32"}}},"ErrorApi":{"properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"},"details":{"type":"object","additionalProperties":{"type":"string"}}}},"MatchStateResponse":{"properties":{"matchId":{"type":"string"},"publicState":{"$ref":"#/components/schemas/PublicGameState"},"privateState":{"$ref":"#/components/schemas/PrivatePlayerState"}}},"PrivateHandCard":{"properties":{"instanceId":{"type":"string"},"cardId":{"type":"string"},"name":{"type":"string"},"supertype":{"type":"string"}}},"PrivatePlayerState":{"properties":{"playerId":{"type":"string","format":"uuid"},"hand":{"type":"array","items":{"$ref":"#/components/schemas/PrivateHandCard"}},"deckCount":{"type":"integer","format":"int32"},"discardCount":{"type":"integer","format":"int32"},"prizes":{"type":"array","items":{"$ref":"#/components/schemas/PrizeSlot"}}}},"PrizeSlot":{"properties":{"slot":{"type":"integer","format":"int32"},"known":{"type":"boolean"},"cardId":{"type":"string"}}},"PublicGameState":{"properties":{"matchId":{"type":"string","format":"uuid"},"status":{"type":"string"},"phase":{"type":"string"},"turnNumber":{"type":"integer","format":"int32"},"currentPlayerId":{"type":"string","format":"uuid"},"firstPlayerId":{"type":"string","format":"uuid"},"players":{"type":"array","items":{"$ref":"#/components/schemas/PublicPlayerState"}}}},"PublicPlayerState":{"properties":{"playerId":{"type":"string","format":"uuid"},"side":{"type":"string"},"activePokemon":{"$ref":"#/components/schemas/PublicPokemonSlot"},"bench":{"type":"array","items":{"$ref":"#/components/schemas/PublicPokemonSlot"}},"prizes":{"type":"array","items":{"type":"string"}}}},"PublicPokemonSlot":{"properties":{"instanceId":{"type":"string"},"cardId":{"type":"string"},"damageCounters":{"type":"integer","format":"int32"},"specialConditions":{"type":"array","items":{"type":"string"}},"attachedCards":{"type":"array","items":{"type":"string"}}}},"CardSearchResponse":{"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/CardSummaryResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalItems":{"type":"integer","format":"int64"}}},"CardSummaryResponse":{"properties":{"id":{"type":"string"},"name":{"type":"string"},"supertype":{"type":"string"},"setCode":{"type":"string"},"number":{"type":"string"},"imageSmallUrl":{"type":"string"}}},"AttackDto":{"properties":{"index":{"type":"integer","format":"int32"},"name":{"type":"string"},"cost":{"type":"array","items":{"type":"string"}},"convertedEnergyCost":{"type":"integer","format":"int32"},"damage":{"type":"string"},"text":{"type":"string"},"baseDamage":{"type":"integer","format":"int32"}}},"CardDetailResponse":{"properties":{"id":{"type":"string"},"name":{"type":"string"},"supertype":{"type":"string"},"subtypes":{"type":"array","items":{"type":"string"}},"setCode":{"type":"string"},"number":{"type":"string"},"imageSmallUrl":{"type":"string"},"imageLargeUrl":{"type":"string"},"rulesText":{"type":"array","items":{"type":"string"}},"hp":{"type":"integer","format":"int32"},"stage":{"type":"string"},"evolvesFrom":{"type":"string"},"types":{"type":"array","items":{"type":"string"}},"attacks":{"type":"array","items":{"$ref":"#/components/schemas/AttackDto"}},"weaknesses":{"type":"array","items":{"$ref":"#/components/schemas/WeaknessDto"}},"resistances":{"type":"array","items":{"$ref":"#/components/schemas/ResistanceDto"}},"retreatCost":{"type":"array","items":{"type":"string"}},"isEx":{"type":"boolean"},"isMega":{"type":"boolean"}}},"ResistanceDto":{"properties":{"type":{"type":"string"},"value":{"type":"string"}}},"WeaknessDto":{"properties":{"type":{"type":"string"},"value":{"type":"string"}}}}}}