# BE/src/main/java/ar/edu/utn/frc/tup/piii/controllers/decks/DeckController.java
package ar.edu.utn.frc.tup.piii.controllers.decks;

import ar.edu.utn.frc.tup.piii.dtos.decks.CreateDeckRequest;
import ar.edu.utn.frc.tup.piii.dtos.decks.DeckResponse;
import ar.edu.utn.frc.tup.piii.dtos.decks.DeckValidationResponse;
import ar.edu.utn.frc.tup.piii.dtos.decks.UpdateDeckRequest;
import ar.edu.utn.frc.tup.piii.dtos.decks.ValidateDeckRequest;
import ar.edu.utn.frc.tup.piii.services.decks.DeckService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/api/decks")
public class DeckController {

    private final DeckService deckService;

    public DeckController(DeckService deckService) {
        this.deckService = deckService;
    }

    @PostMapping
    public ResponseEntity<DeckResponse> createDeck(@RequestBody CreateDeckRequest request) {
        DeckResponse response = deckService.createDeck(request);
        return ResponseEntity.status(HttpStatus.CREATED).body(response);
    }

    @GetMapping("/{id}")
    public ResponseEntity<DeckResponse> getDeck(@PathVariable UUID id) {
        DeckResponse response = deckService.getDeck(id);
        return ResponseEntity.ok(response);
    }

    @PutMapping("/{id}")
    public ResponseEntity<DeckResponse> updateDeck(@PathVariable UUID id, @RequestBody UpdateDeckRequest request) {
        DeckResponse response = deckService.updateDeck(id, request);
        return ResponseEntity.ok(response);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteDeck(@PathVariable UUID id) {
        deckService.deleteDeck(id);
        return ResponseEntity.noContent().build();
    }

    @GetMapping
    public ResponseEntity<List<DeckResponse>> listDecksByPlayer(@RequestParam UUID playerId) {
        List<DeckResponse> decks = deckService.listDecksByPlayer(playerId);
        return ResponseEntity.ok(decks);
    }

    @PostMapping("/{id}/validate")
    public ResponseEntity<DeckValidationResponse> validateDeck(@PathVariable UUID id) {
        DeckValidationResponse response = deckService.validateDeck(id);
        return ResponseEntity.ok(response);
    }

    @PostMapping("/validate")
    public ResponseEntity<DeckValidationResponse> validateCards(@RequestBody ValidateDeckRequest request) {
        DeckValidationResponse response = deckService.validateCards(request);
        return ResponseEntity.ok(response);
    }
}